Send operations to queue
The fog/cloud operations are sent to the message queue in a way that
we can filter them by two properties: topic and key. The first one
represents the message destination which can be 'fog' or 'cloud'.
On the other hand, the key represents the operation itself.

E.g.
`await this.queue.send('fog', 'data.update', { id, data });`

This operations means the we are sending the 'data.update' operation
to the 'fog' topic on the message broker.

Change-Id: I46392d611f259c5bdd7afb6677f8c531403c1ea3
Reviewed-on: https://review.gerrithub.io/c/CESARBR/knot-fog-connector/+/449930
Tested-by: Paulo Serra Filho <ptvsf@cesar.org.br>
Reviewed-by: Paulo Serra Filho <ptvsf@cesar.org.br>