Update patch set 1
Patch Set 1:

I agree on having fog and cloud separated, but since we'll know the cloudType, isn't it possible to have this verification based on it? 
If joi itself is not enough, why not define the schemas for each cloudType and have a simple switch case in loadCloudSettings that would validate the schema based on cloudType?

The main problem I see when not validating here is that if some parameter is not passed, we'll not know what is the issue.

Patch-set: 1
Reviewer: Gerrit User 1015148 <1015148@d5d70762-12d0-45a1-890d-524b12d3f735>