handler: Move AMQP specific operations
Always when the connection with the message broker is re-established
a new channel object is generated. In this context, considering the
message handler receives a channel like that to do some operations
and the instance of it is created with a channel generated before
reconnection, the operations end up failing.

The solution is to respect the separability between layers and avoid
calling a channel object direct in the message handler. So, this patch
moves some of the operations to the AMQP component in a way the channel
is now properly under control.